+++ R/helpers.R
@@ -175,3 +175,382 @@ generate_caption <- function(attributes) {
   # Return the generated caption
   return(caption)
 }
+
+#' Augment Cumulative Sum
+#'
+#' @family Utility Functions
+#' @author Steven P. Sanderson II, MPH
+#' @description This function augments a data frame by adding cumulative sum
+#' columns for specified variables.
+#'
+#' @details The function takes a data frame and a column name (or names) and
+#' computes the cumulative sum for each specified column, starting from an
+#' initial value. If the column names are not provided, it will throw an error.
+#'
+#' @param .data A data frame to augment.
+#' @param .value A column name or names for which to compute the cumulative sum.
+#' @param .names Optional. A character vector of names for the new cumulative
+#' sum columns. Defaults to "auto", which generates names based on the original
+#' column names.
+#' @param .initial_value A numeric value to start the cumulative sum from.
+#' Defaults to 0.
+#'
+#' @return A tibble with the original data and additional columns containing
+#' the cumulative sums.
+#'
+#' @examples
+#' df <- data.frame(x = 1:5, y = 6:10)
+#' std_cum_sum_augment(df, .value = x)
+#' std_cum_sum_augment(df, .value = y, .names = c("cumsum_y"))
+#'
+#' @name std_cum_sum_augment
+NULL
+#' @rdname std_cum_sum_augment
+#' @export
+
+std_cum_sum_augment <- function(.data,
+                                .value,
+                                .names = "auto",
+                                .initial_value = 0) {
+  column_expr <- rlang::enquo(.value)
+
+  if (rlang::quo_is_missing(column_expr)) {
+    rlang::abort("std_cum_sum_augment(.value) is missing.", use_cli_format = TRUE)
+  }
+
+  col_nms <- names(tidyselect::eval_select(column_expr, .data))
+
+  make_call <- function(col) {
+    rlang::expr(!!.initial_value + cumsum(!!rlang::sym(col)))
+  }
+
+  grid <- expand.grid(
+    col = col_nms,
+    stringsAsFactors = FALSE
+  )
+
+  calls <- purrr::pmap(list(grid$col), make_call)
+
+  if (any(.names == "auto")) {
+    newname <- paste0("cum_sum_", grid$col)
+  } else {
+    newname <- as.list(.names)
+  }
+
+  calls <- purrr::set_names(calls, newname)
+
+  ret <- dplyr::as_tibble(dplyr::mutate(.data, !!!calls))
+
+  return(ret)
+}
+
+#' Augment Cumulative Product
+#'
+#' @family Utility Functions
+#' @author Steven P. Sanderson II, MPH
+#' @description This function augments a data frame by adding cumulative product
+#' columns for specified variables.
+#'
+#' @details The function takes a data frame and a column name (or names) and
+#' computes the cumulative product for each specified column, starting from an
+#' initial value. If the column names are not provided, it will throw an error.
+#'
+#' @param .data A data frame to augment.
+#' @param .value A column name or names for which to compute the cumulative product.
+#' @param .names Optional. A character vector of names for the new cumulative
+#' product columns. Defaults to "auto", which generates names based on the original
+#' column names.
+#' @param .initial_value A numeric value to start the cumulative product from.
+#' Defaults to 1.
+#'
+#' @return A tibble with the original data and additional columns containing the
+#' cumulative products.
+#'
+#' @examples
+#' df <- data.frame(x = 1:5, y = 6:10)
+#' std_cum_prod_augment(df, .value = x)
+#' std_cum_prod_augment(df, .value = y, .names = c("cumprod_y"))
+#'
+#' @name std_cum_prod_augment
+NULL
+#' @rdname std_cum_prod_augment
+#' @export
+
+std_cum_prod_augment <- function(.data,
+                                 .value,
+                                 .names = "auto",
+                                 .initial_value = 1) {
+  column_expr <- rlang::enquo(.value)
+
+  if (rlang::quo_is_missing(column_expr)) {
+    stop(call. = FALSE, "std_cum_prod_augment(.value) is missing.")
+  }
+
+  col_nms <- names(tidyselect::eval_select(column_expr, .data))
+
+  make_call <- function(col) {
+    rlang::expr(!!.initial_value * cumprod(1 + !!rlang::sym(col)))
+  }
+
+  grid <- expand.grid(
+    col = col_nms,
+    stringsAsFactors = FALSE
+  )
+
+  calls <- purrr::pmap(.l = list(grid$col), make_call)
+
+  if (any(.names == "auto")) {
+    newname <- paste0("cum_prod_", grid$col)
+  } else {
+    newname <- as.list(.names)
+  }
+
+  calls <- purrr::set_names(calls, newname)
+
+  ret <- dplyr::as_tibble(dplyr::mutate(.data, !!!calls))
+
+  return(ret)
+}
+
+#' Augment Cumulative Minimum
+#'
+#' @family Utility Functions
+#' @author Steven P. Sanderson II, MPH
+#' @description This function augments a data frame by adding cumulative minimum
+#' columns for specified variables.
+#'
+#' @details The function takes a data frame and a column name (or names) and
+#' computes the cumulative minimum for each specified column, starting from an
+#' initial value. If the column names are not provided, it will throw an error.
+#'
+#' @param .data A data frame to augment.
+#' @param .value A column name or names for which to compute the cumulative minimum.
+#' @param .names Optional. A character vector of names for the new cumulative
+#' minimum columns. Defaults to "auto", which generates names based on the
+#' original column names.
+#' @param .initial_value A numeric value to start the cumulative minimum from.
+#' Defaults to 0.
+#'
+#' @return A tibble with the original data and additional columns containing
+#' the cumulative minimums.
+#'
+#' @examples
+#' df <- data.frame(x = c(5, 3, 8, 1, 4), y = c(10, 7, 6, 12, 5))
+#' std_cum_min_augment(df, .value = x)
+#' std_cum_min_augment(df, .value = y, .names = c("cummin_y"))
+#'
+#' @name std_cum_min_augment
+NULL
+#' @rdname std_cum_min_augment
+#' @export
+#'
+std_cum_min_augment <- function(.data,
+                                .value,
+                                .names = "auto",
+                                .initial_value = 0) {
+  column_expr <- rlang::enquo(.value)
+
+  if (rlang::quo_is_missing(column_expr)) {
+    rlang::abort("std_cum_min_augment(.value) is missing.", use_cli_format = TRUE)
+  }
+
+  col_nms <- names(tidyselect::eval_select(column_expr, .data))
+
+  make_call <- function(col) {
+    rlang::expr(!!.initial_value + cummin(!!rlang::sym(col)))
+  }
+
+  grid <- expand.grid(
+    col = col_nms,
+    stringsAsFactors = FALSE
+  )
+
+  calls <- purrr::pmap(list(grid$col), make_call)
+
+  if (any(.names == "auto")) {
+    newname <- paste0("cum_min_", grid$col)
+  } else {
+    newname <- as.list(.names)
+  }
+
+  calls <- purrr::set_names(calls, newname)
+
+  ret <- dplyr::as_tibble(dplyr::mutate(.data, !!!calls))
+
+  return(ret)
+}
+
+#' Augment Cumulative Maximum
+#'
+#' @family Utility Functions
+#' @author Steven P. Sanderson II, MPH
+#' @description This function augments a data frame by adding cumulative maximum
+#' columns for specified variables.
+#'
+#' @details The function takes a data frame and a column name (or names) and
+#' computes the cumulative maximum for each specified column, starting from an
+#' initial value. If the column names are not provided, it will throw an error.
+#'
+#' @param .data A data frame to augment.
+#' @param .value A column name or names for which to compute the cumulative maximum.
+#' @param .names Optional. A character vector of names for the new cumulative
+#' maximum columns. Defaults to "auto", which generates names based on the
+#' original column names.
+#' @param .initial_value A numeric value to start the cumulative maximum from.
+#' Defaults to 0.
+#'
+#' @return A tibble with the original data and additional columns containing the
+#' cumulative maximums.
+#'
+#' @examples
+#' df <- data.frame(x = c(1, 3, 2, 5, 4), y = c(10, 7, 6, 12, 5))
+#' std_cum_max_augment(df, .value = x)
+#' std_cum_max_augment(df, .value = y, .names = c("cummax_y"))
+#'
+#' @name std_cum_max_augment
+NULL
+#' @rdname std_cum_max_augment
+#' @export
+#'
+std_cum_max_augment <- function(.data,
+                                .value,
+                                .names = "auto",
+                                .initial_value = 0) {
+  column_expr <- rlang::enquo(.value)
+
+  if (rlang::quo_is_missing(column_expr)) {
+    rlang::abort("std_cum_max_augment(.value) is missing.", use_cli_format = TRUE)
+  }
+
+  col_nms <- names(tidyselect::eval_select(column_expr, .data))
+
+  make_call <- function(col) {
+    rlang::expr(!!.initial_value + cummax(!!rlang::sym(col)))
+  }
+
+  grid <- expand.grid(
+    col = col_nms,
+    stringsAsFactors = FALSE
+  )
+
+  calls <- purrr::pmap(list(grid$col), make_call)
+
+  if (any(.names == "auto")) {
+    newname <- paste0("cum_max_", grid$col)
+  } else {
+    newname <- as.list(.names)
+  }
+
+  calls <- purrr::set_names(calls, newname)
+
+  ret <- dplyr::as_tibble(dplyr::mutate(.data, !!!calls))
+
+  return(ret)
+}
+
+#' Augment Cumulative Sum
+#'
+#' @family Utility Functions
+#' @author Steven P. Sanderson II, MPH
+#' @description This function augments a data frame by adding cumulative mean
+#' columns for specified variables.
+#'
+#' @details The function takes a data frame and a column name (or names) and
+#' computes the cumulative mean for each specified column, starting from an
+#' initial value. If the column names are not provided, it will throw an error.
+#'
+#' @param .data A data frame to augment.
+#' @param .value A column name or names for which to compute the cumulative mean.
+#' @param .names Optional. A character vector of names for the new cumulative
+#' mean columns. Defaults to "auto", which generates names based on the original
+#' column names.
+#' @param .initial_value A numeric value to start the cumulative mean from.
+#' Defaults to 0.
+#'
+#' @return A tibble with the original data and additional columns containing the
+#' cumulative means.
+#'
+#' @examples
+#' df <- data.frame(x = c(1, 2, 3, 4, 5), y = c(10, 20, 30, 40, 50))
+#' std_cum_mean_augment(df, .value = x)
+#' std_cum_mean_augment(df, .value = y, .names = c("cummean_y"))
+#'
+#' @name std_cum_mean_augment
+NULL
+#' @rdname std_cum_mean_augment
+#' @export
+#'
+std_cum_mean_augment <- function(.data,
+                                 .value,
+                                 .names = "auto",
+                                 .initial_value = 0) {
+  column_expr <- rlang::enquo(.value)
+
+  if (rlang::quo_is_missing(column_expr)) {
+    rlang::abort("std_cum_mean_augment(.value) is missing.", use_cli_format = TRUE)
+  }
+
+  col_nms <- names(tidyselect::eval_select(column_expr, .data))
+
+  make_call <- function(col) {
+    rlang::expr(!!.initial_value + cmean(!!rlang::sym(col)))
+  }
+
+  grid <- expand.grid(
+    col = col_nms,
+    stringsAsFactors = FALSE
+  )
+
+  calls <- purrr::pmap(list(grid$col), make_call)
+
+  if (any(.names == "auto")) {
+    newname <- paste0("cum_mean_", grid$col)
+  } else {
+    newname <- as.list(.names)
+  }
+
+  calls <- purrr::set_names(calls, newname)
+
+  ret <- dplyr::as_tibble(dplyr::mutate(.data, !!!calls))
+
+  return(ret)
+}
+
+#' Get Attributes
+#'
+#' @family Utility Functions
+#'
+#' @author Steven P. Sanderson II, MPH
+#'
+#' @details This function retrieves the attributes of a given R object,
+#' excluding the row.names attribute.
+#'
+#' @description The `get_attributes` function takes an R object as input and
+#' returns its attributes, omitting the row.names attribute.
+#'
+#' @param .data An R object from which attributes are to be extracted.
+#'
+#' @examples
+#' get_attributes(rw30())
+#' get_attributes(iris)
+#' get_attributes(mtcars)
+#'
+#' @return A list of attributes of the input R object, excluding row.names.
+#'
+#' @name get_attributes
+NULL
+#'
+#' @rdname get_attributes
+#'
+#' @export
+
+get_attributes <- function(.data){
+
+  atb <- attributes(.data)
+
+  # drop row.names
+  atb <- atb[!names(atb) %in% c("row.names")]
+
+  # Return
+  return(atb)
+}